swap/wrap to support babashka
