+++ src/clj_uuid/core.clj
@@ -1,0 +1,1065 @@
+(ns clj-uuid.core
+  (:refer-clojure :exclude [== uuid? max < > =])
+  (:require [clojure.core :as clojure]
+            [clj-uuid
+             [constants :refer :all]
+             [util      :as util]
+             [bitmop    :as bitmop]
+             [clock     :as clock]
+             [node      :as node]
+             [random    :as random]])
+  (:import [java.io       ByteArrayOutputStream
+                          ObjectOutputStream]
+           [java.lang IllegalArgumentException]
+           [java.net      URI URL]
+           [java.nio      ByteBuffer]
+           [java.security MessageDigest]
+           [java.util     UUID Date]))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Leach-Salz UUID Representation     [RFC4122:4.1.2 "LAYOUT AND BYTE ORDER"] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;;
+;; The string representation of A Leach-Salz UUID has the format:
+;;
+;;                                          clock-seq-and-reserved
+;;                                time-mid  | clock-seq-low
+;;                                |         | |
+;;                       6ba7b810-9dad-11d1-80b4-00c04fd430c8
+;;                       |             |         |
+;;                       ` time-low    |         ` node
+;;                                     ` time-high-and-version
+;;
+;;
+;; Each field is treated as integer and has its value printed as a zero-filled
+;; hexadecimal digit string with the most significant digit first (unsigned,
+;; big-endian).
+;;
+;; 0                   1                   2                   3
+;;  0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;; |                        %uuid_time-low                         |
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;; |       %uuid_time-mid          |  %uuid_time-high-and-version  |
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;; |clk-seq-hi-res | clock-seq-low |         %uuid_node (0-1)      |
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;; |                         %uuid_node (2-5)                      |
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;;
+;;
+;;  The following table enumerates a slot/type/value correspondence:
+;;
+;;   SLOT       SIZE   TYPE        BYTE-ARRAY
+;;  ----------------------------------------------------------------------
+;;  time-low       4   ub32     [<BYTE> <BYTE> <BYTE> <BYTE>]
+;;  time-mid       2   ub16     [<BYTE> <BYTE>]
+;;  time-high      2   ub16     [<BYTE> <BYTE>]
+;;  clock-high     1    ub8     [<BYTE>]
+;;  clock-low      1    ub8     [<BYTE>]
+;;  node           6   ub48     [<BYTE> <BYTE> <BYTE> <BYTE> <BYTE> <BYTE>]
+;;
+;; Or, as 8-bit bytes mapping into 128-bit unsigned integer values:
+;;
+;;  (0 7)   (8 15)  (16 23) (24 31)  ;; time-low
+;;  (32 39) (40 47)                  ;; time-mid
+;;  (48 55) (56 63)                  ;; time-high-and-version
+;;
+;;  (64 71)                          ;; clock-seq-and-reserved
+;;  (72 79)                          ;; clock-seq-low
+;;  (80 87)   (88 95)   (96 103)     ;;
+;;  (104 111) (112 119) (120 127)    ;; node
+;;
+;; This has been updated with additional layouts.  See RFC9562:5.7 and
+;; the description of v7 UUIDs, below.
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; UUID Variant                                                 [RFC9562:4.1] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;;
+;; The variant indicates the layout of the UUID. The UUID specification
+;; covers one particular variant. Other variants are reserved or exist
+;; for backward compatibility reasons (e.g., for values assigned before
+;; the UUID specification was produced). An example of a UUID that is a
+;; different variant is the null UUID, which is a UUID that has all 128
+;; bits set to zero.
+;;
+;; In the canonical representation:
+;;
+;;    xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx
+;;
+;; the most significant bits of N indicate the variant (depending on the
+;; variant one, two, or three bits are used). The principal variants covered
+;; by the RFC9562 are indicated by the two most significant bits of N being 1 0
+;; (i.e., the hexadecimal N will always be 8, 9, A, or B).  As seen below,
+;; there are two additional variants, currently used for special "sentinal"
+;; UUID's as defined below.
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; UUID Version                                                 [RFC9562:4.2] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;;
+;; The Leach-Salz UUID variant has five defined versions. In the canonical
+;; representation:
+;;
+;;    xxxxxxxx-xxxx-Mxxx-xxxx-xxxxxxxxxxxx
+;;
+;; the four bits of M indicates the UUID version (i.e., the hexadecimal
+;; digit M will be either 1, 2, 3, 4, 5, 6, 7, or 8).
+
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; The NULL (variant 0) UUID                                    [RFC9562:5.9] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(def ^:const +null+
+  "The NULL UUID is a special form of sentinel UUID that is specified to have
+   all 128 bits set to zero."
+  #uuid "00000000-0000-0000-0000-000000000000")
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; The MAX (variant 7) UUID                                    [RFC9562:5.10] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(def ^:const +max+
+  "The MAX UUID is a special form of sentinel UUID that is specified to have
+   all 128 bits set to one."
+  #uuid "FFFFFFFF-FFFF-FFFF-FFFF-FFFFFFFFFFFF")
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Well-Known UUIDs                 [RFC4122:Appendix-C "SOME NAMESPACE IDs"] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+;; The following UUID's are the canonical top-level namespace identifiers
+;; defined in RFC9562 Appendix C.
+
+(def ^:const +namespace-dns+  #uuid "6ba7b810-9dad-11d1-80b4-00c04fd430c8")
+(def ^:const +namespace-url+  #uuid "6ba7b811-9dad-11d1-80b4-00c04fd430c8")
+(def ^:const +namespace-oid+  #uuid "6ba7b812-9dad-11d1-80b4-00c04fd430c8")
+(def ^:const +namespace-x500+ #uuid "6ba7b814-9dad-11d1-80b4-00c04fd430c8")
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Monotonic Clock (guaranteed always increasing value for time)
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn monotonic-time
+ "Return a monotonic timestamp (guaranteed always increasing) based on
+ the number of 100-nanosecond intervals elapsed since the adoption of
+ the Gregorian calendar in the West, 12:00am Friday October 15, 1582 UTC."
+ []
+ (clock/monotonic-time))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; UUID Protocols
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defprotocol UUIDNameBytes
+  "A mechanism intended for user-level extension that defines the
+  decoding rules for the local-part representation of arbitrary
+  Clojure / Java Objects when used for computing namespaced
+  identifiers."
+
+  (as-byte-array [x]
+    "Extract a byte serialization that represents the 'name' of x,
+    typically unique within a given namespace."))
+
+(defprotocol UUIDable
+  "A UUIDable object directly represents a UUID.  Examples of things which
+  might be conceptually 'uuidable' include string representation of a
+  UUID in canonical hex format, or an appropriate URN URI."
+
+  (as-uuid   ^java.util.UUID [x]
+    "Coerce the value 'x' to a UUID.")
+
+  (uuidable?                 [x]
+    "Return 'true' if 'x' can be coerced to UUID."))
+
+;; (defprotocol UUIDRfc9562
+;;   "A protocol that abstracts an unique identifier as described by
+;;   IETF RFC9562 <http://www.ietf.org/rfc/rfc9562.txt>. A UUID
+;;   represents a 128-bit value, however there are variant encoding
+;;   layouts used to assign and interpret information encoded in
+;;   those bits.  This is a protocol for  _variant 2_ (*Leach-Salz*)
+;;   UUID's."
+
+;;   (hash-code                     [uuid])
+;;   (null?                         [uuid])
+;;   (max?                         [uuid])
+;;   (uuid?                         [x])
+;;   (uuid=                         [x y])
+;;   (uuid<                         [x y])
+;;   (uuid>                         [x y])
+;;   (get-word-high                 [uuid])
+;;   (get-word-low                  [uuid])
+;;   (get-version                   [uuid])
+;;   (get-variant                   [uuid])
+;;   (get-time-low                  [uuid])
+;;   (get-time-mid                  [uuid])
+;;   (get-time-high                 [uuid])
+;;   (get-clk-high                  [uuid])
+;;   (get-clk-low                   [uuid])
+;;   (get-clk-seq                   [uuid])
+;;   (get-node-id                   [uuid])
+;;   (get-timestamp                 [uuid])
+;;   (get-instant   ^java.util.Date [uuid])
+;;   (get-unix-time                 [uuid])
+;;   (to-byte-array                 [uuid])
+;;   (to-string     ^String         [uuid])
+;;   (to-hex-string ^String         [uuid])
+;;   (to-urn-string ^String         [uuid])
+;;   (to-uri        ^java.net.URI   [uuid]))
+
+(defprotocol UUIDRfc9562
+  "A protocol that abstracts an unique identifier as described by
+  IETF RFC9562 <http://www.ietf.org/rfc/rfc9562.txt>. A UUID
+  represents a 128-bit value, however there are variant encoding
+  layouts used to assign and interpret information encoded in
+  those bits.  This is a protocol for  _variant 2_ (*Leach-Salz*)
+  UUID's."
+
+  (hash-code                     [uuid]
+    "Return a suitable 64-bit hash value for `uuid`.  Extend with
+    specialized hash computation.")
+
+  (null?                         [uuid]
+    "Return `true` only if `uuid` has all 128 bits set to zero and is
+    therefore equal to the null UUID, 00000000-0000-0000-0000-000000000000.")
+
+  (max?                         [uuid]
+    "Return `true` only if `uuid` has all 128 bits set and is
+    therefore equal to the maximum UUID, ffffffff-ffff-ffff-ffff-ffffffffffff.")
+
+  (uuid?                         [x]
+    "Return `true` if `x` implements an RFC9562 unique identifier.")
+
+  (uuid=                         [x y]
+    "Directly compare two UUID's for = relation based on the equality
+    semantics defined by [RFC4122:3 RULES FOR LEXICAL EQUIVALENCE].
+    See: `clj-uuid/=`")
+
+  (uuid<                         [x y]
+    "Directly compare two UUID's for < relation based on the ordinality
+    semantics defined by [RFC4122:3 RULES FOR LEXICAL EQUIVALENCE].
+    See: `clj-uuid/<`")
+
+  (uuid>                         [x y]
+    "Directly compare two UUID's for > relation based on the ordinality
+    semantics defined by [RFC4122:3 RULES FOR LEXICAL EQUIVALENCE].
+    See: `clj-uuid/>`")
+
+  (get-word-high                 [uuid]
+    "Return the most significant 64 bits of UUID's 128 bit value.")
+
+  (get-word-low                  [uuid]
+    "Return the least significant 64 bits of UUID's 128 bit value.")
+
+  (get-version                   [uuid]
+    "Return the version number associated with this UUID.  The version
+    field contains a value which describes the nature of the UUID.  There
+    are five versions of Leach-Salz UUID, plus the null and max UUIDs:
+
+    0x0   Null
+    0x1   Time based
+    0x2   DCE security with POSIX UID
+    0x3   Namespaced, deterministic (MD5 Digest)
+    0x4   Cryptographic random
+    0x5   Namespaced, deterministic (SHA1 Digest)
+    0x6   Time based, lexically ordered
+    0x7   POSIX Time based, lexically ordered, cryptographically secure
+    0x8   User Customizable
+    0xF   Max
+
+    In the canonical representation, xxxxxxxx-xxxx-Mxxx-xxxx-xxxxxxxxxxxx,
+    the four bits of M indicate the UUID version (i.e., the hexadecimal M
+    will be either 1, 2, 3, 4, 5, 6, 7, or 8).")
+
+  (get-variant                   [uuid]
+    "Return the variant number associated with this UUID.  The variant field
+    contains a value which identifies the layout of the UUID.  The bit-layout
+    implemented by this protocol supports UUID's with a variant value of 0x2,
+    which indicates Leach-Salz layout.  Defined UUID variant values are:
+
+    0x0   Null
+    0x2   Leach-Salz
+    0x6   Microsoft
+    0x7   Max
+
+    In the canonical representation, xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx,
+    the most significant bits of N indicate the variant (depending on the
+    variant one, two, or three bits are used). The variant covered by RFC9562
+    is indicated by the two most significant bits of N being 1 0 (i.e., the
+    hexadecimal N will always be 8, 9, A, or B).")
+
+  (get-time-low                  [uuid]
+    "Return the 32 bit unsigned value that represents the `time-low` field
+    of the `timestamp` associated with this UUID.")
+
+  (get-time-mid                  [uuid]
+    "Return the 16 bit unsigned value that represents the `time-mid` field
+    of the `timestamp` associated with this UUID.")
+
+  (get-time-high                 [uuid]
+    "Return the 16 bit unsigned value that represents the `time-high` field
+    of the `timestamp` multiplexed with the `version` of this UUID.")
+
+  (get-clk-high                  [uuid]
+    "Return the 8 bit unsigned value that represents the most significant
+    byte of the `clk-seq` multiplexed with the `variant` of this UUID.")
+
+  (get-clk-low                   [uuid]
+    "Return the 8 bit unsigned value that represents the least significant
+    byte of the `clk-seq` associated with this UUID.")
+
+  (get-clk-seq                   [uuid]
+    "Return the clock-sequence number associated with this UUID. For time-based
+    (v1, v6) UUID's the 'clock-sequence' value is a somewhat counter-intuitively
+    named seed-value that is used to reduce the potential that duplicate UUID's
+    might be generated under unusual situations, such as if the system hardware
+    clock is set backward in time or if, despite all efforts otherwise, a
+    duplicate node-id happens to be generated. This value is initialized to
+    a random 16-bit number once per lifetime of the system.  For
+    non-gregorian-time-based (v3, v4, v5, v7, v8, squuid) UUID's, always
+    returns `nil`.")
+
+  (get-node-id                   [uuid]
+    "Return the 48 bit unsigned value that represents the spatially unique
+    node identifier associated with this UUID.")
+
+  (get-timestamp                 [uuid]
+    "Return the time of UUID creation.  For Gregorian time-based (v1,
+    v6) UUID's, this is 60 bit unsigned value that represents a
+    temporally unique timestamp associated with this UUID.  The result
+    encodes the number of 100 nanosecond intervals since the adoption of
+    the Gregorian calendar.  For v7 UUID's this encodes the more common
+    unix time in milliseconds since midnight, January 1, 1970 UTC.  For
+    non-time-based (v3, v4, v5, v8, squuid) UUID's, always returns
+    `nil`.")
+
+  (get-instant   ^java.util.Date [uuid]
+    "For time-based (v1, v6, v7) UUID's, return a java.util.Date
+    object that represents the system time at which this UUID was
+    generated. NOTE: the returned value may not necessarily be
+    temporally unique. For non-time-based
+    (v3, v4, v5, v8, squuid) UUID's, always returns `nil`.")
+
+  (get-unix-time                 [uuid]
+    "For time-based (v1, v6, v7) UUIDs return the timestamp portion in
+    approximately milliseconds since the Unix epoch 1970-01-01T00:00:00.000Z.
+    For non-time-based (v3, v4, v5, v8, squuid) UUID's, always returns `nil`.")
+
+  (to-byte-array                 [uuid]
+    "Return an array of 16 bytes that represents `uuid` as a decomposed
+    octet serialization encoded in most-significant-byte first order.")
+
+  (to-string     ^String         [uuid]
+    "Return a String object that represents `uuid` in the canonical
+    36 character hex-string format:
+
+        xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx")
+
+  (to-hex-string ^String         [uuid]
+    "Return a String object that represents `uuid` as the 32 hexadecimal
+    characters directly encodong the UUID's 128 bit value:
+
+        xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx")
+
+  (to-urn-string ^String         [uuid]
+    "Return a String object that represents `uuid` as a the string
+    serialization of the URN URI based on the canonical 36 character
+    hex-string representation:
+
+        urn:uuid:xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx")
+
+  (to-uri        ^java.net.URI   [uuid]
+    "Return the unique URN URI associated with this UUID."))
+
+;; For backwards compatibility
+
+(def UUIDRfc4122 UUIDRfc9562)
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; RFC9562 Unique Identifier extended java.util.UUID
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(extend-type UUID
+
+  UUIDable
+
+  (as-uuid   [u] u)
+  (uuidable? [_] true)
+
+  UUIDRfc9562
+
+  (uuid? ^boolean [_] true)
+
+  (uuid= ^boolean [^UUID x ^UUID y]
+    (.equals x y))
+
+  (uuid< ^boolean [^UUID x ^UUID y]
+    (let [xh (.getMostSignificantBits x)
+          yh (.getMostSignificantBits y)]
+      (or (clojure/< xh yh)
+          (and (clojure/= xh yh) (clojure/< (.getLeastSignificantBits x)
+                                            (.getLeastSignificantBits y))))))
+
+  (uuid> ^boolean [^UUID x ^UUID y]
+    (let [xh (.getMostSignificantBits x)
+          yh (.getMostSignificantBits y)]
+      (or (clojure/> xh yh)
+          (and (clojure/= xh yh) (clojure/> (.getLeastSignificantBits x)
+                                            (.getLeastSignificantBits y))))))
+
+  (get-word-high ^long [uuid]
+    (.getMostSignificantBits uuid))
+
+  (get-word-low ^long [uuid]
+    (.getLeastSignificantBits uuid))
+
+  (null? ^boolean [uuid]
+    (clojure/= 0 (.getMostSignificantBits uuid) (.getLeastSignificantBits uuid)))
+
+  (max? ^boolean [uuid]
+    (uuid= uuid +max+))
+
+  ;; two putLong calls (one per word) instead of bitmop's
+  ;; 16-iteration loop of ldb+sb8 per byte.
+  (to-byte-array ^bytes [uuid]
+    (let [arr (byte-array 16)]
+      (bitmop/long->bytes (.getMostSignificantBits  uuid) arr 0)
+      (bitmop/long->bytes (.getLeastSignificantBits uuid) arr 8)
+      arr))
+
+  (hash-code ^long [uuid]
+    (long (.hashCode uuid)))
+
+  (get-version ^int [uuid]
+    (.version uuid))
+
+  (get-variant ^int [uuid]
+    (.variant uuid))
+
+  (to-string [uuid]
+    (.toString uuid))
+
+  (to-urn-string [uuid]
+    (str "urn:uuid:" (.toString uuid)))
+
+  ;; buf-hex renders all 32 hex chars via a single StringBuilder
+  ;; from a ByteBuffer, instead of two separate hex() calls + string concat.
+
+  (to-hex-string [uuid]
+    (bitmop/buf-hex (bitmop/uuid->buf uuid)))
+
+  (to-uri [uuid]
+    (URI/create (to-urn-string uuid)))
+
+  (get-time-low ^long [uuid]
+    (let [msb (.getMostSignificantBits uuid)]
+      (if (clojure/= 6 (get-version uuid))
+        (bitmop/ldb #=(bitmop/mask 16 0) msb)
+        (bitmop/ldb #=(bitmop/mask 32 0) (bit-shift-right msb 32)))))
+
+  (get-time-mid ^long [uuid]
+    (bitmop/ldb #=(bitmop/mask 16 16)
+      (.getMostSignificantBits uuid)))
+
+  (get-time-high ^long [uuid]
+    (let [msb (.getMostSignificantBits uuid)]
+      (if (clojure/= 6 (get-version uuid))
+        (bitmop/ldb #=(bitmop/mask 32 0) (bit-shift-right msb 32))
+        (bitmop/ldb #=(bitmop/mask 16 0) msb))))
+
+  (get-clk-low ^long [uuid]
+    (bitmop/ldb #=(bitmop/mask 8 0)
+      (bit-shift-right (.getLeastSignificantBits uuid) 56)))
+
+  (get-clk-high ^long [uuid]
+    (bitmop/ldb #=(bitmop/mask 8 48)
+      (.getLeastSignificantBits uuid)))
+
+  (get-clk-seq ^short [uuid]
+    (when (#{1 6} (.version uuid))
+      (bitmop/ldb #=(bitmop/mask 14 48) (.getLeastSignificantBits uuid))))
+
+  (get-node-id ^long [uuid]
+    (bitmop/ldb #=(bitmop/mask 48 0)
+      (.getLeastSignificantBits uuid)))
+
+  (get-timestamp ^long [uuid]
+    (case (.version uuid)
+      1 (.timestamp uuid)
+      6 (bit-or (bitmop/ldb #=(bitmop/mask 12 0)
+                            (.getMostSignificantBits uuid))
+                (bit-shift-left (get-time-mid uuid) 12)
+                (bit-shift-left (get-time-high uuid) 28))
+      7 (bitmop/ldb #=(bitmop/mask 48 16) (.getMostSignificantBits uuid))
+      nil))
+
+  (get-unix-time ^long [uuid]
+    (case (.version uuid)
+      (1 6) (clock/posix-time (get-timestamp uuid))
+      7     (get-timestamp uuid)
+      nil))
+
+  (get-instant [uuid]
+    (when-let [ts (get-unix-time uuid)]
+      (Date. (long ts))))
+
+  UUIDNameBytes
+
+  (as-byte-array
+    ^bytes
+    [this]
+    (to-byte-array this)))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; V0 UUID Constructor                                          [RFC9562:5.9] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn null
+  "Generates the v0 (null) UUID, 00000000-0000-0000-0000-000000000000."
+  ^java.util.UUID
+  []
+  +null+)
+
+(defn v0
+  "Generates the v0 (null) UUID, 00000000-0000-0000-0000-000000000000."
+  ^java.util.UUID
+  []
+  +null+)
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; The MAX UUID Constructor                                    [RFC9562:5.10] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn max
+  "Generates the v15 (maximum) UUID, ffffffff-ffff-ffff-ffff-ffffffffffff."
+  ^java.util.UUID
+  []
+  +max+)
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; V1, V6 (time-coded) UUID Constructors                   [RFC9562:5.1, 5.6] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+;; Concatenate the UUID version with a unique per-node identifier (the
+;; MAC address of the computer that is generating the UUID, or securely
+;; generated once-per-session random identifier) and with a monotonic
+;; timestamp based on the number of 100-nanosecond intervals since the
+;; adoption of the Gregorian calendar in the West, 12:00am Friday
+;; October 15, 1582 UTC.
+
+(let [^java.util.concurrent.atomic.AtomicLong packed clock/-gregorian-packed-
+      subcounter-max clock/+subcounter-resolution+
+      v1-lsb        (long node/+v1-lsb+)
+      v6-lsb        (long node/+v6-lsb+)]
+
+  (defn v1
+    "Generate a v1 (time-based) unique identifier, guaranteed to be unique
+  and thread-safe regardless of clock precision or degree of concurrency.
+  Creation of v1 UUID's does not require any call to a cryptographic
+  generator and can be accomplished much more efficiently than v3, v4, v5, v7,
+  or squuid's.  A v1 UUID reveals both the identity of the computer that
+  generated the UUID and the time at which it did so.  Its uniqueness across
+  computers is guaranteed as long as MAC addresses are not duplicated."
+    ^java.util.UUID
+    []
+    (loop []
+      (let [current  (.get packed)
+            millis   (unsigned-bit-shift-right current 14)
+            time-now (System/currentTimeMillis)]
+        (cond
+          (clojure/< millis time-now)
+          (let [next (bit-shift-left time-now 14)]
+            (if (.compareAndSet packed current next)
+              (let [ts  (+ 100103040000000000
+                           (* (+ 2208988800000 time-now) 10000))
+                    msb (bit-or
+                          (bit-shift-left (bit-and ts 0xFFFFFFFF) 32)
+                          (bit-shift-left (bit-and (unsigned-bit-shift-right ts 32) 0xFFFF) 16)
+                          0x1000
+                          (bit-and (unsigned-bit-shift-right ts 48) 0xFFF))]
+                (UUID. msb v1-lsb))
+              (recur)))
+
+          (clojure/> millis time-now)
+          (recur)
+
+          true
+          (let [seqid (bit-and current 0x3FFF)
+                tt    (inc seqid)]
+            (if (<= tt subcounter-max)
+              (let [next (bit-or (bit-shift-left time-now 14) tt)]
+                (if (.compareAndSet packed current next)
+                  (let [ts  (+ tt 100103040000000000
+                               (* (+ 2208988800000 time-now) 10000))
+                        msb (bit-or
+                              (bit-shift-left (bit-and ts 0xFFFFFFFF) 32)
+                              (bit-shift-left (bit-and (unsigned-bit-shift-right ts 32) 0xFFFF) 16)
+                              0x1000
+                              (bit-and (unsigned-bit-shift-right ts 48) 0xFFF))]
+                    (UUID. msb v1-lsb))
+                  (recur)))
+              (recur)))))))
+
+  (defn v6
+    "Generate a v6 (time-based), LEXICALLY SORTABLE, unique identifier,
+  v6 is a field-compatible version of v1, reordered for improved DB
+  locality.  Creation of v6 UUID's does not require any call to a
+  cryptographic generator and can be accomplished much more efficiently
+  than v3, v4, v5, v7, or squuid's.  A v6 UUID uses a cryptographically
+  secure, hard to guess random node id. It DOES NOT reveal the identity
+  of the computer on which it was created."
+    ^java.util.UUID
+    []
+    (loop []
+      (let [current  (.get packed)
+            millis   (unsigned-bit-shift-right current 14)
+            time-now (System/currentTimeMillis)]
+        (cond
+          (clojure/< millis time-now)
+          (let [next (bit-shift-left time-now 14)]
+            (if (.compareAndSet packed current next)
+              (let [ts  (+ 100103040000000000
+                           (* (+ 2208988800000 time-now) 10000))
+                    msb (bit-or
+                          (bit-shift-left (bit-and (unsigned-bit-shift-right ts 28) 0xFFFFFFFF) 32)
+                          (bit-shift-left (bit-and (unsigned-bit-shift-right ts 12) 0xFFFF) 16)
+                          0x6000
+                          (bit-and ts 0xFFF))]
+                (UUID. msb v6-lsb))
+              (recur)))
+
+          (clojure/> millis time-now)
+          (recur)
+
+          true
+          (let [seqid (bit-and current 0x3FFF)
+                tt    (inc seqid)]
+            (if (<= tt subcounter-max)
+              (let [next (bit-or (bit-shift-left time-now 14) tt)]
+                (if (.compareAndSet packed current next)
+                  (let [ts  (+ tt 100103040000000000
+                               (* (+ 2208988800000 time-now) 10000))
+                        msb (bit-or
+                              (bit-shift-left (bit-and (unsigned-bit-shift-right ts 28) 0xFFFFFFFF) 32)
+                              (bit-shift-left (bit-and (unsigned-bit-shift-right ts 12) 0xFFFF) 16)
+                              0x6000
+                              (bit-and ts 0xFFF))]
+                    (UUID. msb v6-lsb))
+                  (recur)))
+              (recur))))))))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; V7 (crypto secure, posix time-coded) UUID Constructor        [RFC9562:5.7] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;;
+;;  0                   1                   2                   3
+;;  0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;; |                           unix_ts_ms                          |
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;; |          unix_ts_ms           |  ver  |       rand_a          |
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;; |var|                        rand_b                             |
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+;; |                            rand_b                             |
+;; +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+
+;;  The following table enumerates a slot/type/value correspondence:
+;;
+;;   SLOT       SIZE      Description
+;;  ----------------------------------------------------------------------
+;;  unix_ts_ms   6 bytes  POSIX millis timestamp
+;;  ver          4 bits   version, set to 0b0111 (7)
+;;  rand_a      12 bits   randomly initialized subcounter
+;;  var          2 bits   variant, set to 0b10 (2)
+;;  rand_b      62 bits   cryptographically secure pseudorandom data
+
+(defn v7
+  "Generate a v7 unix time-based, LEXICALLY SORTABLE UUID with monotonic
+  counter and cryptographically secure random portion and POSIX time encoding.
+  As such, creation of v7 UUIDs may be significantly slower, but have improved
+  entropy chararacteristics compared to v1 or v6 UUIDs."
+  ^java.util.UUID
+  []
+  (let [^clj_uuid.clock.State state (clock/monotonic-unix-time-and-random-counter)
+        time            (bitmop/ldb #=(bitmop/mask 48  0) (.millis state))
+        ver-and-counter (bitmop/dpb #=(bitmop/mask 4  12) (.seqid state) 0x7)
+        msb             (bit-or ver-and-counter (bit-shift-left time 16))
+        lsb             (bitmop/dpb #=(bitmop/mask 2 62) (random/long) 0x2)]
+    (UUID. msb lsb)))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; V7nc (non-cryptographic V7) UUID Constructor                              ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(let [^ThreadLocal v7nc-tl
+      (ThreadLocal/withInitial
+        (reify java.util.function.Supplier
+          (get [_] (long-array 3))))]
+  (defn v7nc
+    "Generate a v7 UUID using non-cryptographic randomness (ThreadLocalRandom).
+  Produces valid RFC 9562 v7 UUIDs with the same timestamp/version/variant
+  structure as v7, but uses java.util.concurrent.ThreadLocalRandom instead
+  of SecureRandom and a per-thread monotonic counter instead of a global
+  AtomicLong.  This trades cryptographic unguessability for significantly
+  higher throughput -- comparable to JUG's TimeBasedEpochGenerator.
+
+  The per-thread counter uses rand_a (12 bits) as a fixed seed per
+  millisecond and rand_b (62 bits) as a strictly increasing counter,
+  giving 2^62 UUIDs per millisecond per thread before overflow."
+    ^java.util.UUID
+    []
+    (let [^longs state (.get v7nc-tl)
+          ^java.util.concurrent.ThreadLocalRandom tlr
+          (java.util.concurrent.ThreadLocalRandom/current)]
+      (loop []
+        (let [time-now (System/currentTimeMillis)
+              last-ms  (aget state 0)]
+          (cond
+            (clojure/> time-now last-ms)
+            (let [lsb-ctr (bit-and (.nextLong tlr) 0x3FFFFFFFFFFFFFFF)
+                  msb     (bit-or (bit-shift-left (bit-and time-now 0xFFFFFFFFFFFF) 16)
+                                  (bit-or 0x7000 (bit-and (.nextLong tlr) 0xFFF)))]
+              (aset state 0 time-now)
+              (aset state 1 msb)
+              (aset state 2 lsb-ctr)
+              (UUID. msb (bit-or lsb-ctr #=(bit-shift-left 2 62))))
+
+            (clojure/< time-now last-ms)
+            (recur)
+
+            true
+            (let [lsb-ctr (bit-and (unchecked-inc (aget state 2)) 0x3FFFFFFFFFFFFFFF)]
+              (aset state 2 lsb-ctr)
+              (UUID. (aget state 1) (bit-or lsb-ctr #=(bit-shift-left 2 62))))))))))
+
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; V4 (random) UUID Constructor                                 [RFC9562:5.4] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn v4
+  "Generate a v4 (random) UUID.  Uses default JVM implementation.  If two
+  arguments, lsb and msb (both long) are provided, then construct a valid,
+  properly formatted v4 UUID based on those values.  So, for example the
+  following UUID, created from all zero bits, is indeed distinct from the
+  null UUID:
+
+      (v4)
+       => #uuid \"dcf0035f-ea29-4d1c-b52e-4ea499c6323e\"
+
+      (v4 0 0)
+       => #uuid \"00000000-0000-4000-8000-000000000000\"
+
+      (null)
+       => #uuid \"00000000-0000-0000-0000-000000000000\""
+  (^java.util.UUID []
+    (UUID/randomUUID))
+  (^java.util.UUID [msb lsb]
+    (UUID.
+      (bitmop/dpb #=(bitmop/mask 4 12) msb 0x4)
+      (bitmop/dpb #=(bitmop/mask 2 62) lsb 0x2))))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; v8 (custom) UUID Constructor                  [RFC9562:5.8: UUID Version 8];;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn v8
+  "Generate a v8 custom UUID with up to 122 bits of user data."
+  ^java.util.UUID
+  [^long msb ^long lsb]
+  (UUID.
+   (bitmop/dpb #=(bitmop/mask 4 12) msb 0x8)
+   (bitmop/dpb #=(bitmop/mask 2 62) lsb 0x2)))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; SQUUID (sequential) UUID Constructor
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn squuid
+  "Generate a SQUUID (sequential, random) unique identifier.  SQUUID's
+  are a nonstandard variation on v4 (random) UUIDs that have the
+  desirable property that they increase sequentially over time as well
+  as encode retrievably the posix time at which they were generated.
+  Splits and reassembles a v4 UUID to merge current POSIX
+  time (seconds since 12:00am January 1, 1970 UTC) with the most
+  significant 32 bits of the UUID."
+  ^java.util.UUID
+  []
+  (let [uuid (v4)
+        secs (clock/posix-time)
+        lsb  (get-word-low  uuid)
+        msb  (get-word-high uuid)
+        timed-msb (bit-or (bit-shift-left secs 32)
+                    (bit-and +ub32-mask+ msb))]
+    (UUID. timed-msb lsb)))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; "Local-Part" Representation
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+;; The following represent a default set of local-part encoding rules.  As a
+;; default, a plain byte-array will be passed through unchanged and a
+;; generic java.lang.Object is represented by the bytes of its serialization.
+;; Strings are represented using UTF8 encoding.  URL's are digested as the
+;; UTF bytes of their string representation.
+
+(def ^:private ByteArray (class (byte-array 0)))
+
+(extend-protocol UUIDNameBytes
+
+  java.lang.Object
+  (as-byte-array ^bytes [this]
+    (if (instance? ByteArray this)
+      this
+      (let [baos (ByteArrayOutputStream.)
+            oos  (ObjectOutputStream. baos)]
+        (.writeObject oos this)
+        (.close oos)
+        (.toByteArray baos))))
+
+  java.lang.String
+  (as-byte-array ^bytes [this]
+    (util/compile-if (util/java6?)
+      (.getBytes this)
+      (.getBytes this java.nio.charset.StandardCharsets/UTF_8)))
+
+  java.net.URL
+  (as-byte-array ^bytes [this]
+    (as-byte-array (.toString this)))
+
+  nil
+  (as-byte-array [x]
+    (throw (IllegalArgumentException. (format "%s cannot be converted to byte array." x)))))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Digest Instance
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(def ^:private -md5-
+  (ThreadLocal/withInitial
+    (reify java.util.function.Supplier
+      (get [_] (MessageDigest/getInstance "MD5")))))
+
+(def ^:private -sha1-
+  (ThreadLocal/withInitial
+    (reify java.util.function.Supplier
+      (get [_] (MessageDigest/getInstance "SHA1")))))
+
+(defn- make-digest
+  ^java.security.MessageDigest
+  [^String designator]
+  (let [^MessageDigest md
+        (case designator
+          "MD5"  (.get ^ThreadLocal -md5-)
+          "SHA1" (.get ^ThreadLocal -sha1-)
+          (MessageDigest/getInstance designator))]
+    (.reset md)
+    md))
+
+(defn- digest-bytes
+  ^bytes
+  [^String kind ^bytes ns-bytes ^bytes local-bytes]
+  (let [m (make-digest kind)]
+    (.update m ns-bytes)
+    (.digest m local-bytes)))
+
+(defn- build-digested-uuid
+  ^java.util.UUID
+  [^long version ^bytes arr]
+  {:pre [(or (clojure/= version 3) (clojure/= version 5))]}
+  (let [msb (bitmop/bytes->long arr 0)
+        lsb (bitmop/bytes->long arr 8)]
+    (UUID.
+     (bitmop/dpb #=(bitmop/mask 4 12) msb version)
+     (bitmop/dpb #=(bitmop/mask 2 62) lsb 0x2))))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Namespaced UUIDs                                        [RFC9562:5.3, 5.5] ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(let [^ThreadLocal md5-tl  -md5-
+      ^ThreadLocal sha1-tl -sha1-
+      ^ThreadLocal ns-buf-tl (ThreadLocal/withInitial
+                               (reify java.util.function.Supplier
+                                 (get [_] (ByteBuffer/allocate 16))))]
+
+  (defn v3
+    "Generate a v3 (name based, MD5 hash) UUID. 'context' must be UUIDable.
+  v3 identifiers are intended for generating UUID's from names that are
+  drawn from, and unique within, some namespace.  The concept of name and
+  namespace should be broadly construed, and not limited to textual names.
+  The requirements for a v3 UUID are as follows:
+
+  * v3 UUID's generated at different times from the same name in the same
+    namespace MUST be equal.
+
+  * v3 UUID's generated from two different names in the same namespace
+    SHOULD be distinct with a high degree of certainty.
+
+  * v3 UUID's generated from the same name in two different namespaces
+    SHOULD be distinct with a high degree of certainty.
+
+  * If two v3 UUID's are equal, then there is a high degree of certainty
+    that they were generated from the same name in the same namespace."
+    ^java.util.UUID
+    [context local-part]
+    (let [^MessageDigest md (.get md5-tl)
+          _                 (.reset md)
+          ^UUID ns-uuid     (as-uuid context)
+          ^ByteBuffer nsbuf (.get ns-buf-tl)
+          _                 (.putLong nsbuf 0 (.getMostSignificantBits ns-uuid))
+          _                 (.putLong nsbuf 8 (.getLeastSignificantBits ns-uuid))
+          _                 (.update md (.array nsbuf))
+          digest            (.digest md ^bytes (as-byte-array local-part))
+          ^ByteBuffer dbuf  (ByteBuffer/wrap digest)
+          msb               (bit-or (bit-and (.getLong dbuf 0)
+                                     #=(bit-not #=(bitmop/mask 4 12)))
+                              0x3000)
+          lsb               (bit-or (bit-and (.getLong dbuf 8)
+                                     #=(bit-not #=(bitmop/mask 2 62)))
+                              #=(bit-shift-left 2 62))]
+      (UUID. msb lsb)))
+
+  (defn v5
+    "Generate a v5 (name based, SHA1 hash) UUID. 'context' must be UUIDable.
+  v5 identifiers are intended for generating UUID's from names that are
+  drawn from, and unique within, some namespace.  The concept of name and
+  namespace should be broadly construed, and not limited to textual names.
+  The requirements for a v5 UUID are as follows:
+
+  * v5 UUID's generated at different times from the same name in the same
+    namespace MUST be equal.
+
+  * v5 UUID's generated from two different names in the same namespace
+    SHOULD be distinct with a high degree of certainty.
+
+  * v5 UUID's generated from the same name in two different namespaces
+    SHOULD be distinct with a high degree of certainty.
+
+  * If two v5 UUID's are equal, then there is a high degree of certainty
+    that they were generated from the same name in the same namespace."
+    ^java.util.UUID
+    [context local-part]
+    (let [^MessageDigest md (.get sha1-tl)
+          _                 (.reset md)
+          ^UUID ns-uuid     (as-uuid context)
+          ^ByteBuffer nsbuf (.get ns-buf-tl)
+          _                 (.putLong nsbuf 0 (.getMostSignificantBits ns-uuid))
+          _                 (.putLong nsbuf 8 (.getLeastSignificantBits ns-uuid))
+          _                 (.update md (.array nsbuf))
+          digest            (.digest md ^bytes (as-byte-array local-part))
+          ^ByteBuffer dbuf  (ByteBuffer/wrap digest)
+          msb               (bit-or (bit-and (.getLong dbuf 0)
+                                     #=(bit-not #=(bitmop/mask 4 12)))
+                              0x5000)
+          lsb               (bit-or (bit-and (.getLong dbuf 8)
+                                     #=(bit-not #=(bitmop/mask 2 62)))
+                              #=(bit-shift-left 2 62))]
+      (UUID. msb lsb))))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Predicates
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn- compare-many
+  [f x y more]
+  (if (f x y)
+    (if (next more)
+      (recur f y (first more) (next more))
+      (f y (first more)))
+    false))
+
+(defn =
+  "Directly compare two or more UUIDs for = relation based on the
+  equality semantics defined by [RFC4122:3 RULES FOR LEXICAL
+  EQUIVALENCE]."
+  ([_] true)
+  ([x y]
+   (uuid= x y))
+  ([x y & more]
+   (compare-many uuid= x y more)))
+
+(defn >
+  "Directly compare two or more UUIDs for > relation based on the
+  ordinality semantics defined by [RFC4122:3 RULES FOR LEXICAL
+  EQUIVALENCE]."
+  ([_] true)
+  ([x y]
+   (uuid> x y))
+  ([x y & more]
+   (compare-many uuid> x y more)))
+
+(defn <
+  "Directly compare two or more UUIDs for < relation based on the
+  ordinality semantics defined by [RFC4122:3 RULES FOR LEXICAL
+  EQUIVALENCE]."
+  ([_] true)
+  ([x y]
+   (uuid< x y))
+  ([x y & more]
+   (compare-many uuid< x y more)))
+
+(defn uuid-string? [str]
+  (and (string? str)
+       (some? (re-matches uuid-regex str))))
+
+(defn uuid-urn-string? [str]
+  (and (string? str)
+       (some? (re-matches urn-regex str))))
+
+(defn uuid-vec? [v]
+  (and (clojure/= (count v) 16)
+       (every? #(and (integer? %) (>= -128  % 127)) v)))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; UUID Polymorphism
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn- str->uuid [s]
+  (cond
+    (uuid-string?     s) (UUID/fromString s)
+    (uuid-urn-string? s) (UUID/fromString (subs s 9))
+    :else                (throw (IllegalArgumentException. (format "Invalid UUID: %s" s)))))
+
+(extend-protocol UUIDRfc9562
+  Object
+  (uuid? [x] false)
+
+  nil
+  (uuid? [_] false))
+
+(extend-protocol UUIDable
+  (Class/forName "[B") ; byte array
+  (as-uuid [^bytes ba]
+    (let [bb (ByteBuffer/wrap ba)]
+      (UUID. (.getLong bb 0) (.getLong bb 8))))
+  (uuidable? [^bytes ba]
+    (clojure/= 16 (alength ^bytes ba)))
+
+  String
+  (uuidable? ^boolean [s]
+    (or
+     (uuid-string?     s)
+     (uuid-urn-string? s)))
+  (as-uuid [s]
+    (str->uuid s))
+
+  URI
+  (uuidable? ^boolean [u]
+    (uuid-urn-string? (str u)))
+  (as-uuid [u]
+    (str->uuid (str u)))
+
+  Object
+  (uuidable? ^boolean [_]
+    false)
+  (as-uuid [x]
+    (throw (IllegalArgumentException. (format "%s Cannot be coerced to UUID." x))))
+
+  nil
+  (as-uuid [x]
+    (throw (IllegalArgumentException. (format "%s cannot be coerced to UUID." x))))
+  (uuidable? ^boolean [_] false))
